requirements.txt
# === Core Frameworks ===
langchain
langchain-community
langchain-google-genai
langchain-huggingface


# === Vector Search & Embeddings ===
faiss-cpu
sentence-transformers
numpy

# === Web Framework & UI ===
streamlit

# === PDF & Data Utilities ===
pypdf
python-dotenv

# === Google Gemini SDK ===
google-generativeai

# === Optional Utilities ===
tqdm
requests

# === Hugging Face Support ===
huggingface-hub

# === General Purpose Helpers ===
typing-extensions
